requirements-offline.txt
Deprecated==1.3.1
Flask==3.1.3
requests==2.32.5
waitress==3.0.2
bosdyn-orbit==5.1.1
